fix: cast group-count expression to size_t for std::min (WASM build)
On the emscripten/em++ (WASM) CI job, size_t is 'unsigned long' while
uint64_t is 'unsigned long long', so the mixed-type std::min call in the
parallel CONSTRUCT export path fails template argument deduction.  Cast
the computed group count to size_t explicitly; on all 64-bit platforms
this is a no-op.

+++ src/engine/ExportQueryExecutionTrees.cpp
@@ -894,8 +894,9 @@ ExportQueryExecutionTrees::constructQueryResultToStream(
   size_t maxGroups = std::max<size_t>(1, 4 * numThreads);
   size_t rowsPerGroup = bufferMemory / (numThreads * triplesPerRow * 128);
   rowsPerGroup = std::max<size_t>(1, rowsPerGroup);
-  maxGroups =
-      std::min(maxGroups, (totalRows + rowsPerGroup - 1) / rowsPerGroup);
+  maxGroups = std::min(
+      maxGroups,
+      static_cast<size_t>((totalRows + rowsPerGroup - 1) / rowsPerGroup));
   maxGroups = std::max<size_t>(1, maxGroups);
   std::vector<std::vector<TableWithRange>> groups =
       splitBlocksIntoGroups(blocks, maxGroups);
